+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 1.1.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Refactor to one single IExpandableElement adapter. @ksuess
+- Add IBlockSearchableText adapter for accordion block. @ksuess
+- Add IBlockSearchableText adapter for teaser block. @ksuess
 
 
 1.1.3 (2023-04-22)