+++ CHANGES.rst
@@ -5,9 +5,10 @@ Changelog
 1.1.4 (unreleased)
 ------------------
 
-- Refactor to one single IExpandableElement adapter. @ksuess
-- Add IBlockSearchableText adapter for accordion block. @ksuess
-- Add IBlockSearchableText adapter for teaser block. @ksuess
+- Get elastic client getter from collective.elastic.ingest. [ksuess]
+- Refactor to one single IExpandableElement adapter. [ksuess]
+- Add IBlockSearchableText adapter for accordion block. [ksuess]
+- Add IBlockSearchableText adapter for teaser block. [ksuess]
 
 
 1.1.3 (2023-04-22)